Add Result class to avoid the overhead of Promise when possible
Using Promises/async seems to be significantly slower than returning the
result immediately even if the async method immediately returns the
result. The overhead seems to be worse in Firefox than in Chrome, but
even in Chrome it is significantly slower. Using Observable is slightly
better than Promise, but it is still much slower.

For RemoteBTree the nodes returned by fetchNode might often be available
immediately (e.g. because they might be cached), so to avoid the
overhead in those cases add a new Result class that can either hold the
result directly or a Promise for the result. Using this is almost as
fast as returning the result directly (if it is available immediately).
And even if Promises are used it is still faster than using only
Promises, because intermediate Promises are avoided.

The code is not as nice as always using async/await. It is probably
comparable to using Observable, so it should be a good compromise.