Implement moving patches to page group pages
Remove newPageNumbers and pageNumberToTransactionId from IndexPage, they
are not needed: newPageNumbers is kind of redundant to
pageGroupNumberToTransactionId and pageNumberToTransactionId can just be
in the page group pages (it is fine to always write the page group page
if an actual page is written).

Do some related refactoring, fixes and improvements.