Remove entries count from inner btree pages
This avoids modifying all the pages up to the root every time an entry
changes. It also reduces the complexity.

Storing the entries count in the inner pages does not seem very useful
after all, in particular considering the way indexes are planned to work
(bitmaps will be used to store multiple "index entries" in one entry, so
the raw entry count does not mean much).