Improve test performance
Replace usages of expect() with the "low level" assert(). assert() is
much faster than expect(), in particular when very many
expects/assertions happen in a loop.

These changes reduce the test runtime from about 5.1s to about 3.1s.

The downside is, that the output in case af failures is worse, but those
should hopefully be rare.