Implement an optional common prefix in pageEntries (to save space)
Automatically detect if there is a common prefix between the first three
entries and if so, then use that as potential prefix for all entries of
the page (entries without the prefix can still also be added as well).

Do some other refactorings and improvements.

Remove the unused debugPageEntries() function.